+++ annotation/src/main/java/org/cbioportal/genome_nexus/annotation/service/internal/VEPIsoformOverrideService.java
@@ -1,15 +1,13 @@
 package org.cbioportal.genome_nexus.annotation.service.internal;
 
 import org.cbioportal.genome_nexus.annotation.domain.IsoformOverride;
+import org.cbioportal.genome_nexus.annotation.domain.IsoformOverrideRepoFactory;
 import org.cbioportal.genome_nexus.annotation.domain.IsoformOverrideRepository;
-import org.cbioportal.genome_nexus.annotation.domain.internal.IsoformOverrideRepositoryImpl;
+import org.cbioportal.genome_nexus.annotation.domain.internal.IsoformOverrideRepoFactoryImpl;
 import org.cbioportal.genome_nexus.annotation.service.IsoformOverrideService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
-
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * @author Selcuk Onur Sumer
@@ -17,48 +15,18 @@ import java.util.Map;
 @Service
 public class VEPIsoformOverrideService implements IsoformOverrideService
 {
-    private String overrideURIs;
-
-    private Map<String, IsoformOverrideRepository> overrideRepositories;
+    private final IsoformOverrideRepoFactory repoFactory;
 
     @Autowired
-    public VEPIsoformOverrideService(@Value("${vep.overrides_uris}") String overrideURIs)
+    public VEPIsoformOverrideService(
+        @Qualifier("defaultIsoformOverrideRepoFactory") IsoformOverrideRepoFactory repoFactory)
     {
-        this.overrideURIs = overrideURIs;
-
-        Map<String, String> resources = parseOverrideURIsString(overrideURIs);
-        this.overrideRepositories = new HashMap<>();
-
-        // Create a repository for each resource URI
-        for (String key: resources.keySet())
-        {
-            this.overrideRepositories.put(key, new IsoformOverrideRepositoryImpl(resources.get(key)));
-        }
-    }
-
-    private Map<String, String> parseOverrideURIsString(String overrideURIs)
-    {
-        Map<String, String> overrideResources = new HashMap<>();
-
-        for (String pair: overrideURIs.split(","))
-        {
-            String parts[] = pair.split(":");
-
-            if (parts.length > 1)
-            {
-                String key = parts[0];
-                String value = parts[1];
-
-                overrideResources.put(key, value);
-            }
-        }
-
-        return overrideResources;
+        this.repoFactory = repoFactory;
     }
 
     public IsoformOverride getIsoformOverride(String source, String id)
     {
-        IsoformOverrideRepository repository = this.overrideRepositories.get(source);
+        IsoformOverrideRepository repository = this.repoFactory.getRepository(source);
 
         if (repository != null)
         {
@@ -68,15 +36,5 @@ public class VEPIsoformOverrideService implements IsoformOverrideService
         {
             return null;
         }
-    }
-
-    public String getOverrideURIs()
-    {
-        return overrideURIs;
-    }
-
-    public Map<String, IsoformOverrideRepository> getOverrideRepositories()
-    {
-        return overrideRepositories;
     }
 }
